Move ProxySelector to provider for reuse elsewhere
